Add offsetTransforms property, externalize function.
